+++ src/V1/ServiceBricks.Work.Model/Model/WorkModelConstants.cs
@@ -1,0 +1,13 @@
+namespace ServiceBricks.Work
+{
+    /// <summary>
+    /// Constants for the Work module.
+    /// </summary>
+    public static partial class WorkModelConstants
+    {
+        /// <summary>
+        /// Application settings keys for client API configuration.
+        /// </summary>
+        public const string APPSETTING_CLIENT_APICONFIG = @"ServiceBricks:Work:Client:Api";
+    }
+}