+++ src/V1/ServiceBricks.Work/Model/WorkConstants.cs
@@ -5,9 +5,6 @@ namespace ServiceBricks.Work
     /// </summary>
     public static partial class WorkConstants
     {
-        /// <summary>
-        /// Application settings keys for client API configuration.
-        /// </summary>
-        public const string APPSETTING_CLIENT_APICONFIG = @"ServiceBricks:Work:Client:Api";
+
     }
 }