+++ src/V1/Tests/TestFiles/InMemory/ProcessStubApiClientTests.cs
@@ -1,0 +1,94 @@
+﻿using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using ServiceBricks.Work;
+
+namespace ServiceBricks.Xunit
+{
+    public class ProcessStubTestManager : ProcessTestManager
+    {
+        public class ProcessHttpClientFactory : IHttpClientFactory
+        {
+            private ApiClientTests.CustomGenericHttpClientHandler<ProcessDto> _handler;
+
+            public ProcessHttpClientFactory(ApiClientTests.CustomGenericHttpClientHandler<ProcessDto> handler)
+            {
+                _handler = handler;
+            }
+
+            public HttpClient CreateClient(string name)
+            {
+                return new HttpClient(_handler);
+            }
+        }
+
+        public override IApiClient<ProcessDto> GetClient(IServiceProvider serviceProvider)
+        {
+            var config = new ConfigurationBuilder()
+            .AddInMemoryCollection(new Dictionary<string, string?>
+            {
+                            { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ReturnResponseObject", "false" },
+                            { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":DisableAuthentication", "false" },
+                            { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":TokenUrl", "https://localhost:7000/token" },
+                            { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":BaseServiceUrl", "https://localhost:7000/" },
+            })
+            .Build();
+
+            var apioptions = new OptionsWrapper<ApiOptions>(new ApiOptions() { ReturnResponseObject = false });
+            var apiservice = serviceProvider.GetRequiredService<IProcessApiService>();
+            var controller = new ProcessApiController(apiservice, apioptions);
+            var handler = new ApiClientTests.CustomGenericHttpClientHandler<ProcessDto>(controller);
+            var clientHandlerFactory = new ProcessHttpClientFactory(handler);
+            return new ProcessApiClient(
+                serviceProvider.GetRequiredService<ILoggerFactory>(),
+                clientHandlerFactory,
+                config);
+        }
+
+        public ApiClientTests.CustomGenericHttpClientHandler<ProcessDto> Handler { get; set; }
+
+        public override IApiClient<ProcessDto> GetClientReturnResponse(IServiceProvider serviceProvider)
+        {
+            var config = new ConfigurationBuilder()
+            .AddInMemoryCollection(new Dictionary<string, string?>
+            {
+                            { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":ReturnResponseObject", "true" },
+                            { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":DisableAuthentication", "false" },
+                            { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":TokenUrl", "https://localhost:7000/token" },
+                            { ServiceBricksConstants.APPSETTING_CLIENT_APIOPTIONS + ":BaseServiceUrl", "https://localhost:7000/" },
+            })
+            .Build();
+
+            var apioptions = new OptionsWrapper<ApiOptions>(new ApiOptions() { ReturnResponseObject = true });
+            var apiservice = serviceProvider.GetRequiredService<IProcessApiService>();
+            var controller = new ProcessApiController(apiservice, apioptions);
+            var handler = new ApiClientTests.CustomGenericHttpClientHandler<ProcessDto>(controller);
+            var clientHandlerFactory = new ProcessHttpClientFactory(handler);
+            return new ProcessApiClient(
+                serviceProvider.GetRequiredService<ILoggerFactory>(),
+                clientHandlerFactory,
+                config);
+        }
+    }
+
+    [Collection(ServiceBricks.Xunit.Constants.SERVICEBRICKS_COLLECTION_NAME)]
+    public class StubProcessApiClientTest : ApiClientTest<ProcessDto>
+    {
+        public StubProcessApiClientTest()
+        {
+            SystemManager = ServiceBricksSystemManager.GetSystemManager(typeof(StartupInMemory));
+            TestManager = new ProcessStubTestManager();
+        }
+    }
+
+    [Collection(ServiceBricks.Xunit.Constants.SERVICEBRICKS_COLLECTION_NAME)]
+    public class StubProcessApiClientReturnResponseTests : ApiClientReturnResponseTest<ProcessDto>
+    {
+        public StubProcessApiClientReturnResponseTests()
+        {
+            SystemManager = ServiceBricksSystemManager.GetSystemManager(typeof(StartupInMemory));
+            TestManager = new ProcessStubTestManager();
+        }
+    }
+}